Move certain exceptions from utils to `sipa.model.exceptions`
