+++ sipa/model/gerok/__init__.py
@@ -4,7 +4,7 @@ from ipaddress import IPv4Network
 
 from ..datasource import DataSource, Dormitory
 from . import user
-from sipa.utils.exceptions import InvalidConfiguration
+from sipa.model.exceptions import InvalidConfiguration
 
 
 def init_context(app):
