+++ sipa/model/hss/user.py
@@ -15,7 +15,7 @@ from sipa.model.hss.ldap import HssLdapConnector, change_password
 from sipa.model.hss.schema import Account, IP, AccountStatementLog, TrafficQuota
 from sipa.units import money
 from sipa.utils import argstr
-from sipa.utils.exceptions import InvalidCredentials
+from sipa.model.exceptions import InvalidCredentials
 logger = logging.getLogger(__name__)
